Factor out commonality between client.c & server.c
